🚀 Try wasm implementation
