+++ plugins/web_adapter/lib/dio_web_adapter.dart
@@ -3,22 +3,22 @@ library dio_web_adapter;
 // We separate implementations as `html` and `js_interop`
 // only for convenient diffing.
 
-export 'src/html/adapter.dart';
+// export 'src/html/adapter.dart';
 
-// export 'src/js_interop/adapter.dart';
+export 'src/js_interop/adapter.dart';
 
-export 'src/html/compute.dart';
+// export 'src/html/compute.dart';
 
-// export 'src/js_interop/compute.dart';
+export 'src/js_interop/compute.dart';
 
-export 'src/html/dio_impl.dart';
+// export 'src/html/dio_impl.dart';
 
-// export 'src/js_interop/dio_impl.dart';
+export 'src/js_interop/dio_impl.dart';
 
-export 'src/html/multipart_file.dart';
+// export 'src/html/multipart_file.dart';
 
-// export 'src/js_interop/multipart_file.dart';
+export 'src/js_interop/multipart_file.dart';
 
-export 'src/html/progress_stream.dart';
+// export 'src/html/progress_stream.dart';
 
-// export 'src/js_interop/progress_stream.dart';
+export 'src/js_interop/progress_stream.dart';
