+++ plugins/web_adapter/lib/src/js_interop/compute.dart
@@ -19,19 +19,19 @@
 // The file is intentionally not refactored so that it is easier to keep the
 // compute package up to date with Flutter's implementation.
 
-export '../html/compute.dart';
+// export '../html/compute.dart';
 
-// import 'package:dio/src/compute/compute.dart' as c;
-//
-// /// The dart:html implementation of [c.compute].
-// Future<R> compute<Q, R>(
-//   c.ComputeCallback<Q, R> callback,
-//   Q message, {
-//   String? debugLabel,
-// }) async {
-//   // To avoid blocking the UI immediately for an expensive function call, we
-//   // pump a single frame to allow the framework to complete the current set
-//   // of work.
-//   await null;
-//   return callback(message);
-// }
+import 'package:dio/src/compute/compute.dart' as c;
+
+/// The dart:html implementation of [c.compute].
+Future<R> compute<Q, R>(
+  c.ComputeCallback<Q, R> callback,
+  Q message, {
+  String? debugLabel,
+}) async {
+  // To avoid blocking the UI immediately for an expensive function call, we
+  // pump a single frame to allow the framework to complete the current set
+  // of work.
+  await null;
+  return callback(message);
+}
