+++ plugins/web_adapter/pubspec.yaml
@@ -7,20 +7,20 @@ topics:
   - http
   - network
   - web
-#  - wasm
+  - wasm
 homepage: https://github.com/cfug/dio
 repository: https://github.com/cfug/dio/blob/main/plugins/web_adapter
 issue_tracker: https://github.com/cfug/dio/issues
 
 environment:
-  sdk: '>=2.18.0 <4.0.0'
-#  sdk: '>=3.3.0 <4.0.0'
+#  sdk: '>=2.18.0 <4.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
   dio: ^5.0.0
   http_parser: ^4.0.0
   meta: ^1.5.0
-#  web: '>=0.5.0 <0.6.0'
+  web: '>=0.5.0 <0.6.0'
 
 dev_dependencies:
   lints: any
